Added tool to assign measurement values to grid coodinates.
Satellite data is available for grid coordinates, but oceanic fishing data coordinates do not fall perfectly on grid axes. This tool locates the square grid tile within which a fishing survey falls, then averages the measured values of the tile verteces and assigns this value to the survey coordinate. Note that the user should inspect the variance of the results to identify any tiles whose four vertex values are extremely different. This may be the case, for example, near shorelines and continental shelf borders.